% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HieRFIT_functions.R
\name{ceR}
\alias{ceR}
\title{Certainty Estimation function. Calculates certainty values for each class probability.}
\usage{
ceR(qP, qW)
}
\arguments{
\item{qP}{Observed probability for each class}

\item{qW}{Probability weights estimated by graWeighteR()}
}
\value{
QueCers
}
\description{
Certainty Estimation function. Calculates certainty values for each class probability.
}
