% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HieRFIT_functions.R
\name{ClassProbCalculator}
\alias{ClassProbCalculator}
\title{A function to calculate class scores for all internal and tip node classes.}
\usage{
ClassProbCalculator(tree, nodes_P_all)
}
\arguments{
\item{nodes_P_all}{a table output from CTTraverser() which contains all class probabilities from every node.}
}
\value{
P_path_prod a table for products of ancestor node scores.
}
\description{
A function to calculate class scores for all internal and tip node classes.
}
