% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_engine.R
\name{CreateRef}
\alias{CreateRef}
\title{The main function for creating a reference model.}
\usage{
pbmc.refmod <- CreateRef(Ref = as.matrix(pbmc@data), ClassLabels = pbmc@meta.data$ClusterNames_0.6, TreeFile = pbmc3k_tree)
}
\arguments{
\item{Ref}{Reference data from which class labels will be projected on Query data.}

\item{ClassLabels}{A list of class labels for cells/samples in the ExpData matrix (Ref). Same length as colnames(Ref).}

\item{method}{The model training method, "rf" for random forest, "svmLinear" for support vector machine, "hrf" for hierarchical random forest. Default is "hrf"}

\item{TreeFile}{An input file to create a hierarchical tree for relationship between class labels. Default is null but required if method 'hrf' is chosen.}
}
\description{
The main function for creating a reference model.
}
