% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hieRF_functions.R
\name{CreateTree}
\alias{CreateTree}
\title{A function to create a tree in phylo format.}
\usage{
treeTable <- read.delim("TaxonomyRank.tree.txt", header = F)
}
\arguments{
\item{treeTable}{a data table/data.frame storing class relationships with intermediate class labels. Can be read from a tab separated file.}
}
\description{
A function to create a tree in phylo format.
}
