% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hieRF_functions.R
\name{CTTraverser}
\alias{CTTraverser}
\title{A hierarchical Classifier Tree Traverser function.}
\usage{
CTTraverser(Query, tree, hiemods, thread = NULL)
}
\arguments{
\item{Query}{is the input query data. rows are genes and columns are cells.}

\item{tree}{a tree topology with which hrf model was trained on. 'phylo' format.}

\item{hiemods}{models list from HieRandForest function.}

\item{thread}{number of workers to be used for parallel processing. Default is Null, so serial processing.}
}
\description{
A hierarchical Classifier Tree Traverser function.
}
