% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_engine.R
\name{DataReshaper}
\alias{DataReshaper}
\title{An internal function to prepare a training/test dataset for model generation.}
\usage{
trainingData <- DataReshaper(ExpData = as.matrix(SeuratObject@data), Predictors = genes, ClassLabels = SeuratObject@meta.data$CellTypes)
}
\arguments{
\item{ExpData}{a Normalized expression data matrix, genes in rows and samples in columns.}

\item{Predictors}{the predictor feature list selected by FeatureSelector.}

\item{ClassLabels}{[optional] A list of class labels for cells/samples in the ExpData matrix. Same length as colnames(ExpData).}

\item{alpa}{variation cutoff for filtering data}
}
\description{
An internal function to prepare a training/test dataset for model generation.
}
\keyword{data}
\keyword{preparation}
