% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{DownSizeSeurat}
\alias{DownSizeSeurat}
\title{A function to downsample Seurat object based on cell type identity}
\usage{
pbmc1 <- DownSizeSeurat(SeuratObj = pbmc, IdentityCol = 7)
}
\arguments{
\item{SeuratObj}{a Seurat S4 object.}

\item{IdentityCol}{the column 'number' in the metadata slot showing the cell type identities.}
}
\description{
A function to downsample Seurat object based on cell type identity
}
