% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{EvaluateCertainty}
\alias{EvaluateCertainty}
\title{A function to evaluate performance of HieRFIT with various Certainty thresholds.}
\usage{
EvaluateCertainty(RefSeuObj, IdentityCol, RefMod, Uinter = 20,
  perm_n = 10, samp_n = 100)
}
\arguments{
\item{RefSeuObj}{reference seurat object.}

\item{IdentityCol}{the column 'number' in the metadata slot showing the cell type identities.}

\item{Uinter}{number of intervals to assess certainty threshold.}

\item{perm_n}{# of permutations.}

\item{samp_n}{# samples to draw at each permutation.}
}
\description{
A function to evaluate performance of HieRFIT with various Certainty thresholds.
}
