% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{FeatureSelector}
\alias{FeatureSelector}
\title{A function used internally for selecting genes based on their weight in the top principle components.}
\usage{
FeatureSelector(ExpData, PCs, num, spitPlots = F,
  prefix = "gene.select")
}
\arguments{
\item{ExpData}{an data matrix storing gene expression as genes in rows and samples in columns.}

\item{PCs}{the number of PCs to be looked at when selecting genes.}

\item{num}{the number of genes in total to be included.}

\item{spitPlots}{boolean to generate a pdf for PCA plots. Default is False.}

\item{prefix}{a prefix to separate each run.}
}
\description{
p: pca object, pcs: number of PCs to include, num: number of genes from top and bottom.
Weighted gene picking depending on PC number: Initial PCs give more genes.
For example; the number of genes are taken from PC i is calculated by (pcs-i+1)*(num*2)/(pcs*(pcs+1))
}
\examples{
genes <- as.character(SelectGenesBestLoadings(ExpData = train, prefix = "test.run", pcs = 10, num = 2000))
}
\keyword{PCA}
\keyword{loadings}
\keyword{selection}
