% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FixLab}
\alias{FixLab}
\title{A function to fix class labels.}
\usage{
FixLab(xstring)
}
\arguments{
\item{xstring}{is a list of class labels in character.}
}
\description{
A function to fix class labels.
}
