% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetChildNodeLeafs}
\alias{GetChildNodeLeafs}
\title{A function to retrieve corresponding leafs of the child nodes of a given node.}
\usage{
GetChildNodeLeafs(tree, node)
}
\arguments{
\item{tree}{A tree storing relatinship between the class labels.}

\item{node}{a particular non-terminal node in the tree.}
}
\description{
A function to retrieve corresponding leafs of the child nodes of a given node.
}
