% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HieRFIT_functions.R
\name{graWeighteR}
\alias{graWeighteR}
\title{A function to calculate gravitity center (weight center) of the probability distributions
among classes of the node. By permutation.}
\usage{
graWeighteR(model, QueData)
}
\arguments{
\item{model}{of the node.}

\item{QueData}{same matrix used in PvoteR().}
}
\value{
QueWeights a set of probability weightes per class to be used in asymetrix entropy estimations.
}
\description{
A function to calculate gravitity center (weight center) of the probability distributions
among classes of the node. By permutation.
}
