% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHier_functions.R
\name{HieRandForest}
\alias{HieRandForest}
\title{An internal function to create hierarchical classification models (hiemods) with random forest.}
\usage{
HieRandForest(RefData, ClassLabels, tree, thread = 3, ...)
}
\arguments{
\item{RefData}{a Normalized expression data matrix, genes in rows and samples in columns.}

\item{ClassLabels}{A list of class labels for cells/samples in the Data matrix. Same length as colnames(Data).}

\item{tree}{A tree storing relatinship between the class labels. Default is null. required when mod.meth "hrf" is chosen.}

\item{thread}{number of workers to be used for parallel processing. Default is Null, so serial processing.}
}
\description{
An internal function to create hierarchical classification models (hiemods) with random forest.
}
