% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HieRFIT_functions.R
\name{HieRFIT}
\alias{HieRFIT}
\title{The HieRFIT Class}
\usage{
HieRFIT(Query, refMod, Prior = NULL, xSpecies = NULL, alpha = 0.9)

expRefObj <- get(load("data/exp_refObj.Rdata"))
}
\arguments{
\item{Query}{Query data whose components will be labeled with reference data.}

\item{refMod}{optional input if model exist already. Default is null and generated from scratch. Input can be an caret model object or a .Rdata file.}

\item{Prior}{prior class labels if exist. For cross comparison. Should correspond row order of the Query.}

\item{xSpecies}{optional argument to specify cross species information transfer. Default is null. Possible options are 'rat2mouse', 'mouse2rat', 'mouse2human', human2mouse. With respect to model data.}
}
\description{
The HieRFIT Class

The main function to project reference data on query in order to identify class labels.
}
