% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{hPRF}
\alias{hPRF}
\title{A function for Hierarchical Precision, Recall, and F-measure.}
\usage{
hPRF(tpT, tree, BetaSq = 1)
}
\arguments{
\item{tpT}{PriorPostTable: a table with two columns of which first is Prior and second is Post-prediction.}

\item{tree}{tree topology in phylo format.}
}
\description{
A function for Hierarchical Precision, Recall, and F-measure.
}
