% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{IntSectSize}
\alias{IntSectSize}
\title{A function to determine the size of intersection between ancestors of True class and Predicted class}
\usage{
IntSectSize(t, p, tree)
}
\arguments{
\item{t}{true class}

\item{p}{predicted class}
}
\description{
A function to determine the size of intersection between ancestors of True class and Predicted class
}
