% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_engine.R
\name{Modeller}
\alias{Modeller}
\title{An internal function to collect model training parameters and direct them to model creation.}
\usage{
rf.model <- Modeller(ExpData = as.matrix(pbmc1@data), ClassLabels = pbmc1@meta.data$ClusterNames_0.6)
}
\arguments{
\item{ExpData}{a Normalized expression data matrix, genes in rows and samples in columns.}

\item{ClassLabels}{A list of class labels for cells/samples in the ExpData matrix. Same length as colnames(ExpData).}

\item{mod.meth}{The model training method, "rf" for random forest, "svmLinear" for support vector machine, "hrf" for hierarchical random forest. Default is "rf"}

\item{cv.k}{Fold cross validation. Default is 5.}

\item{tree}{A tree storing relatinship between the class labels. Default is null. required when mod.meth "hrf" is chosen.}

\item{save.int.f}{Boolean to save model. Default is False.}
}
\description{
An internal function to collect model training parameters and direct them to model creation.
}
