% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHier_functions.R
\name{NodeTrainer}
\alias{NodeTrainer}
\title{A function to create a local classifier for a given node.}
\usage{
NodeTrainer(Rdata, tree, node, f_n = 200, tree_n = 500, ...)
}
\arguments{
\item{f_n}{number of features to be included in local classifier.}
}
\description{
A function to create a local classifier for a given node.
}
