% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploting_functions.R
\name{PlotTopo}
\alias{PlotTopo}
\title{This function uses 'ape' package.}
\usage{
PlotTopo(RandTreeSim(LN = sample(5:20, 1)))
}
\arguments{
\item{tree}{a tree object in phylo format.}
}
\description{
This function uses 'ape' package.
}
