% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_engine.R
\name{Predictor}
\alias{Predictor}
\title{The predictor function.}
\usage{
Predictor(model, Query, format = "prob", node = NULL)
}
\arguments{
\item{format}{type of prediction output, "prob" or "resp".}
}
\description{
The predictor function.
}
