% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{prepareDataset}
\alias{prepareDataset}
\title{An internal function to prepare a training/test dataset for model generation.}
\usage{
prepareDataset(ExpData, ClassLabels, PCs, featureGeneSet,
  prefix = "prep.run", ...)
}
\arguments{
\item{ExpData}{a Normalized expression data matrix, genes in rows and samples in columns. Example: .}

\item{ClassLabels}{A list of class labels for cells/samples in the ExpData matrix. Same length as colnames(ExpData).}

\item{PCs}{the number of PCs to be looked at when selecting genes.}

\item{featureGeneSet}{an optional list of genes to be used in the training matrix rather than PC based selection. If this is provided, PC selection is skipped.}

\item{prefix}{a prefix to separate each run.}
}
\description{
An internal function to prepare a training/test dataset for model generation.
}
\examples{
trainingData <- prepareDataset(ExpData = as.matrix(SeuratObject@data), ClassLabels = SeuratObject@meta.data$CellTypes, PCs = 10)
}
\keyword{Training}
\keyword{preparation}
