% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HieRFIT_functions.R
\name{PvoteR}
\alias{PvoteR}
\title{A function to tally votes from the model locally trained.}
\usage{
PvoteR(model, QueData, format = "prob", node = NULL)
}
\arguments{
\item{QueData}{is the prepared data matrix ready to be used in predict}

\item{format}{type of prediction output, "prob" or "resp".}
}
\description{
A function to tally votes from the model locally trained.
}
