% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hieRF_functions.R
\name{RandTreeSim}
\alias{RandTreeSim}
\title{A function to generate a random tree using 'ape' package.
Returns a tree object.}
\usage{
RandTreeSim(LN = 8, furcation = "binary")
}
\arguments{
\item{LN}{The number of leafs to be in the tree. Default is 8.}
}
\description{
A function to generate a random tree using 'ape' package.
Returns a tree object.
}
