% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_engine.R
\docType{class}
\name{RefMod-class}
\alias{RefMod-class}
\alias{RefMod}
\title{Reference class}
\description{
Reference class
}
\section{Slots}{

\describe{
\item{\code{model}}{A list of models to be used.}

\item{\code{tree}}{A hierarchical tree representing class relationships.}

\item{\code{modtype}}{type of the model method used to create object.}
}}

