% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HieRFIT_functions.R
\name{scoR}
\alias{scoR}
\title{The predictor function.}
\usage{
scoR(model, QueData, format = "prob", node = NULL)
}
\arguments{
\item{format}{type of prediction output, "prob" or "resp".}
}
\description{
The predictor function.
}
