% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HieRFIT_functions.R
\name{ScoreEval}
\alias{ScoreEval}
\title{A function for evalating the uncertainty.}
\usage{
ScoreEval(ScoreObs, ProbCert, alpha = 0.9)
}
\arguments{
\item{ScoreObs}{P_path_prod for observed scores}

\item{ProbCert}{Certainty scores.}
}
\description{
A function for evalating the uncertainty.
}
