% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHier_functions.R
\name{SubsetTData}
\alias{SubsetTData}
\title{A function to slide data according to the class hierarchies. And updates the class labels.}
\usage{
SubsetTData(Tdata, tree, node)
}
\description{
A function to slide data according to the class hierarchies. And updates the class labels.
}
