% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_engine.R
\name{SvmWrap}
\alias{SvmWrap}
\title{A wrapper function for random forest.}
\usage{
SvmWrap(ExpData = ExpData, ClassLabels = ClassLabels, prefix, mod.meth,
  train.control, ncor = 5, ...)
}
\arguments{
\item{ExpData}{a Normalized expression data matrix, genes in rows and samples in columns.}

\item{ClassLabels}{A list of class labels for cells/samples in the ExpData matrix. Same length as colnames(ExpData).}

\item{mod.meth}{The model training method, "svmLinear" for support vector machine.}
}
\description{
A wrapper function for random forest.
}
