% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHier_functions.R
\name{FeatureSelector}
\alias{FeatureSelector}
\title{A function used internally for selecting genes based on their weight in the top principle components.}
\usage{
Predictors <- FeatureSelector(Data = as.matrix(SeuratObject@data), PCs = 10, num = 2000)
}
\arguments{
\item{Data}{an data matrix storing gene expression as genes in rows and samples in columns.}

\item{ClassLabels}{A list of class labels for cells/samples in the Data matrix. Same length as colnames(Data).}

\item{PC_n}{the number of PCs to be looked at when selecting genes. Default is 40.}

\item{num}{the number of Predictors (genes) in total to be included. Default is 2000.}

\item{...}{parameters to be passed down to subfunctions such as f_n, tree_n, and PC_n,
for "number of features per local classifier", "number of trees per local classsifier",
and "number of PC space for feature search", respectively.}
}
\description{
p: pca object, pcs: number of PCs to include, num: number of genes from top and bottom.
Weighted gene picking depending on PC number: Initial PCs give more genes.
For example; the number of genes are taken from PC i is calculated by (pcs-i+1)*(num*2)/(pcs*(pcs+1))
}
\keyword{PCA}
\keyword{loadings}
\keyword{selection}
