% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{QuickSeurat}
\alias{QuickSeurat}
\title{A wrapper function for quick data processing with Seurat functions
This function allows to determine highly variable genes and scale their expression,
run PCA, tSNE, and cluster detection.}
\usage{
pbmc <- QuickSeurat(pbmc, scale.only.var=F, PCs=5, perp=20)
}
\arguments{
\item{SeuratObj}{a Seurat S4 object.}

\item{scale.only.var}{a boolean variable to determine whether to scale entire data or only highly variable genes. Default is True.}

\item{PCs}{number of PCs to be included in the downstream analysis}

\item{perp}{perplexity parameter to be passed to RunTSNE function from Seurat.}

\item{vars2reg}{variables to be regressed out when scaling the expression data.}
}
\description{
A wrapper function for quick data processing with Seurat functions
This function allows to determine highly variable genes and scale their expression,
run PCA, tSNE, and cluster detection.
}
\keyword{quick}
\keyword{seurat}
